Minister commands; fix copypasta namespaces
