+++ FrEee.Core.Domain/Gameplay/Commands/Designs/IDesignCommandFactory.cs
@@ -3,13 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FrEee.Gameplay.Commands.Designs;
 using FrEee.Objects.Civilization;
 using FrEee.Objects.Civilization.Orders;
 using FrEee.Objects.GameState;
 using FrEee.Objects.Vehicles;
 
-namespace FrEee.Gameplay.Commands.Orders;
+namespace FrEee.Gameplay.Commands.Designs;
 
 /// <summary>
 /// Builds various types of commands used for managing vehicle designs.
