+++ FrEee.Core.Domain/Gameplay/Commands/ICommandFactory.cs
@@ -12,8 +12,5 @@ namespace FrEee.Gameplay.Commands;
 public interface ICommandFactory
 {
     // TODO: split these out into separate factories for each category and add parameters
-
-    ICommand MinisterToggle();
-
     ICommand Research();
 }
