+++ FrEee.Core.Domain/Gameplay/Commands/Waypoints/IWaypointCommandFactory.cs
@@ -5,16 +5,16 @@ using System.Text;
 using System.Threading.Tasks;
 using FrEee.Objects.Civilization;
 
-namespace FrEee.Gameplay.Commands;
+namespace FrEee.Gameplay.Commands.Waypoints;
 
 /// <summary>
 /// Builds commands to manage waypoints.
 /// </summary>
 public interface IWaypointCommandFactory
 {
-    ICommand<Empire> CreateWaypoint(Waypoint waypoint);
+	ICommand<Empire> CreateWaypoint(Waypoint waypoint);
 
-    ICommand<Waypoint> DeleteWaypoint(Waypoint waypoint);
+	ICommand<Waypoint> DeleteWaypoint(Waypoint waypoint);
 
-    ICommand<Waypoint> HotkeyWaypoint(Waypoint waypoint, int hotkey, bool redirect);
+	ICommand<Waypoint> HotkeyWaypoint(Waypoint waypoint, int hotkey, bool redirect);
 }
