+++ FrEee.Core.Domain/Utility/DIRoot.cs
@@ -3,10 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FrEee.Gameplay.Commands;
+using FrEee.Gameplay.Commands.Designs;
 using FrEee.Gameplay.Commands.Fleets;
 using FrEee.Gameplay.Commands.Messages;
+using FrEee.Gameplay.Commands.Ministers;
+using FrEee.Gameplay.Commands.Notes;
 using FrEee.Gameplay.Commands.Orders;
+using FrEee.Gameplay.Commands.Waypoints;
 using FrEee.Processes;
 using FrEee.Processes.Combat;
 
@@ -41,6 +44,11 @@ public static class DIRoot
 	/// Allows players to manage diplomatic messages.
 	/// </summary>
 	public static IMessageCommandFactory MessageCommands => DI.Get<IMessageCommandFactory>();
+
+	/// <summary>
+	/// Allows players to manage AI ministers.
+	/// </summary>
+	public static IMinisterCommandFactory MinisterCommands => DI.Get<IMinisterCommandFactory>();
 
 	/// <summary>
 	/// Allows players to manage notes.
